FIX: Crocoblock/issues-tracker#3089

+++ includes/post-meta/base-meta-type.php
@@ -28,7 +28,7 @@ abstract class Base_Meta_Type implements
 
 	public function query( $form_id ) {
 		if ( false === $form_id ) {
-			$form_id = jet_fb_live()->form_id;
+			$form_id = (int) jet_fb_live()->form_id;
 		}
 		return Tools::decode_json(
 			get_post_meta(
